Create javascript for library
Create an array to hold the library items in. Write a constructor
function as a template for all of the library books. Create a
function to add new books to the library array. Write a function
that presents all of the book items in cards on the page. Include
an option to change the books status from read to not read, and an
option to remove the book from the page and the array. Create a
button that brings up a form for users to submit more books to the
page.